Add bash_functions script.
bash_functions is sourced by bashrc. It mainly changes cd to behave
as pushd does with functions and aliases.

I also rewrote the setup.py script to be neater and to resolve issue
where you cannot backup to an already existent relationship/
There are some other small fixes in there as well.